implement new task script

+++ routes/tasks_api.js
@@ -10,16 +10,13 @@ const router = express.Router();
 const query = require('../db/queries/tasks');
 const newCategory = require('./helpers/categoryApi.js');
 
-//change
-//delete
-
 router.get('/', (req, res) => {
-  // const userId = req.session.userId;
-  // if (!userId) {
-  //   res.error("Please log in");
-  //   return;
-  // }
-  const userId = 1;
+  const userId = req.session.userID;
+  if (!userId) {
+    res.status(401).json({error: "Please log in"});
+    return;
+  }
+  // const userId = 1;
   query.getAllUserTasks(userId)
     .then(data => {
       res.json(data);
@@ -31,13 +28,14 @@ router.get('/', (req, res) => {
 });
 
 router.post('/create', (req, res) => {
-  if (!req.body.text) {
+  if (!req.body.newTask) {
     res.status(400).json({ error: 'invalid request: no data in POST body' });
     return;
   }
-  // const userId = req.session.userId;
-  const userId = 1;
-  const task = req.body.text;
+  const userId = req.session.userID;
+  console.log(userId)
+  // const userId = 1;
+  const task = req.body.newTask;
 
   newCategory.newCategory(task)
     .then((category) => {
@@ -61,12 +59,12 @@ router.post('/create', (req, res) => {
 });
 
 router.post('/change/:id', (req, res) => {
-  // const userId = req.session.userId;
-  // if (!userId) {
-  //   res.error("Please log in");
-  //   return;
-  // }
-  const userId = 1;
+  const userId = req.session.userID;
+  if (!userId) {
+    res.error("Please log in");
+    return;
+  }
+  // const userId = 1;
   const taskId = req.params.id;
   const newCategory = req.body.newCategory;
   query.editTask(userId, taskId, newCategory)
@@ -81,12 +79,12 @@ router.post('/change/:id', (req, res) => {
 });
 
 router.post('/delete/:id', (req, res) => {
-  // const userId = req.session.userId;
-  // if (!userId) {
-  //   res.error("Please log in");
-  //   return;
-  // }
-  const userId = 1;
+  const userId = req.session.userID;
+  if (!userId) {
+    res.error("Please log in");
+    return;
+  }
+  // const userId = 1;
   const taskId = req.params.id;
   query.deleteTask(userId, taskId)
     .then(() => {
